+++ CHANGELOG.md
@@ -8,7 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-* Nothing
+* add [CodeFactor](https://www.codefactor.io/) badge to main ```README.md```
 
 ### Changed